failed attempt to implement css variable outside of global scope
Essentially, there is a solution in which the non-global scoped css variables can be recognized. This would involve completely converting the tree and then looking through the arbitrary rules for '--' variable declaration/assignments or 'var(--' variable usages, reading through, adding things to unitDict and the tailwind config file, and then recalculate ALL rules with the new unitDict. however, they would not be treated in their relevant scope due to the flattening of the tree. therefore, I will leave it unless it is requested.

Another solution is to recursively look through the tree and find all variable declarations and add things to unit dict and tailwind config before the tailwind conversion.